Crate virtual_dom_rs

virtual-dom-rs provides a virtual dom implementation as well as an html! macro that you can use to generate a virtual dom.

The virtual dom works on both the client and server. On the client we'll render to an HtmlElement, and on the server we render to a String.

Re-exports

pub extern crate web_sys;

pub use crate::html_macro::*;

pub use crate::virtual_node::*;

Modules

Closure
Runtime support for the wasm-bindgen tool

html_macro
The html_macro module exposes an html! macro that is used to generate VirtualNode's that will eventually get rendered into DOM nodes on the client side, or String's if you're on the server side.

prelude
Exports structs and macros that you'll almost always want access to in a virtual-dom powered application

virtual_node
The virtual_node module exposes the VirtualNode struct and methods that power our virtual dom.

Macros

html
A macro which returns a root VirtualNode given some HTML and Rust expressions.

recurse_html

Structs

Comment
The Comment object

Document
The Document object

DomTokenList
The DOMTokenList object

Element
The Element object

Event
The Event object

EventTarget
The EventTarget object

HtmlCollection
The HTMLCollection object

HtmlElement
The HTMLElement object

Node
The Node object

NodeList
The NodeList object

Text
The Text object

Window
The Window object

Enums

Patch
A Patch encodes an operation that modifies a real DOM element.

Traits

View
A trait with common functionality for rendering front-end views.

Functions

diff
Given two VirtualNode's generate Patch's that would turn the old virtual node's real DOM node equivalent into the new VirtualNode's real DOM node equivalent.

patch
Apply all of the patches to our old root node in order to create the new root node that we desire. This is usually used after diffing two virtual nodes.

window