Trait virtual_dom_rs::Closure::convert::RefFromWasmAbi

pub trait RefFromWasmAbi: WasmDescribe where
    <Self::Anchor as Deref>::Target == Self, {
    type Abi: WasmAbi;
    type Anchor: Deref;
    unsafe fn ref_from_abi(js: Self::Abi, extra: &mut dyn Stack) -> Self::Anchor;
}

A trait for anything that can be recovered as some sort of shared reference from the wasm ABI boundary.

This is the shared reference variant of the opposite operation as IntoWasmAbi.

Associated Types

type Abi: WasmAbi

The wasm ABI type references to Self are recovered from.

type Anchor: Deref

The type that holds the reference to Self for the duration of the invocation of the function that has an &Self parameter. This is required to ensure that the lifetimes don't persist beyond one function call, and so that they remain anonymous.

Required Methods

unsafe fn ref_from_abi(js: Self::Abi, extra: &mut dyn Stack) -> Self::Anchor

Recover a Self::Anchor from Self::Abi.

Safety

Same as FromWasmAbi::from_abi.