Struct virtual_dom_rs::Closure::prelude::JsValue

pub struct JsValue { /* fields omitted */ }

Representation of an object owned by JS.

A JsValue doesn't actually live in Rust right now but actually in a table owned by the wasm-bindgen generated JS glue code. Eventually the ownership will transfer into wasm directly and this will likely become more efficient, but for now it may be slightly slow.

Methods

impl JsValue

pub const NULL: JsValue

The null JS value constant.

pub const UNDEFINED: JsValue

The undefined JS value constant.

pub const TRUE: JsValue

The true JS value constant.

pub const FALSE: JsValue

The false JS value constant.

pub fn from_str(s: &str) -> JsValue

Creates a new JS value which is a string.

The utf-8 string provided is copied to the JS heap and the string will be owned by the JS garbage collector.

pub fn from_f64(n: f64) -> JsValue

Creates a new JS value which is a number.

This function creates a JS value representing a number (a heap allocated number) and returns a handle to the JS version of it.

pub fn from_bool(b: bool) -> JsValue

Creates a new JS value which is a boolean.

This function creates a JS object representing a boolean (a heap allocated boolean) and returns a handle to the JS version of it.

pub fn undefined() -> JsValue

Creates a new JS value representing undefined.

pub fn null() -> JsValue

Creates a new JS value representing null.

pub fn symbol(description: Option<&str>) -> JsValue

Creates a new JS symbol with the optional description specified.

This function will invoke the Symbol constructor in JS and return the JS object corresponding to the symbol created.

pub fn as_f64(&self) -> Option<f64>

Returns the f64 value of this JS value if it's an instance of a number.

If this JS value is not an instance of a number then this returns None.

pub fn is_string(&self) -> bool

Tests whether this JS value is a JS string.

pub fn as_string(&self) -> Option<String>

If this JS value is a string value, this function copies the JS string value into wasm linear memory, encoded as UTF-8, and returns it as a Rust String.

To avoid the copying and re-encoding, consider the as_js_string() method instead.

If this JS value is not an instance of a string or if it's not valid utf-8 then this returns None.

pub fn as_bool(&self) -> Option<bool>

Returns the bool value of this JS value if it's an instance of a boolean.

If this JS value is not an instance of a boolean then this returns None.

pub fn is_null(&self) -> bool

Tests whether this JS value is null

pub fn is_undefined(&self) -> bool

Tests whether this JS value is undefined

pub fn is_symbol(&self) -> bool

Tests whether the type of this JS value is symbol

pub fn is_object(&self) -> bool

Tests whether typeof self == "object" && self !== null.

pub fn is_function(&self) -> bool

Tests whether the type of this JS value is function.