Function virtual_dom_rs::patch

pub fn patch<N: Into<Node>>(
    root_node: N, 
    patches: &Vec<Patch>
) -> Result<HashMap<u32, Vec<DynClosure>>, JsValue>

Apply all of the patches to our old root node in order to create the new root node that we desire. This is usually used after diffing two virtual nodes.