Struct virtual_dom_rs::Comment

#[repr(transparent)]pub struct Comment { /* fields omitted */ }

The Comment class.

MDN Documentation

This API requires the following crate features to be activated: Comment

Implementations

impl Comment

pub fn new() -> Result<Comment, JsValue>

The new Comment(..) constructor, creating a new instance of Comment.

MDN Documentation

This API requires the following crate features to be activated: Comment

impl Comment

pub fn new_with_data(data: &str) -> Result<Comment, JsValue>

The new Comment(..) constructor, creating a new instance of Comment.

MDN Documentation

This API requires the following crate features to be activated: Comment