Struct virtual_dom_rs::Document

#[repr(transparent)]

pub struct Document { /* fields omitted */ }

The Document object

MDN Documentation

This API requires the following crate features to be activated: Document

Methods

impl Document

pub fn new() -> Result<Document, JsValue>

The new Document(..) constructor, creating a new instance of Document

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn adopt_node(&self, node: &Node) -> Result<Node, JsValue>

The adoptNode() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn create_comment(&self, data: &str) -> Comment

The createComment() method

MDN Documentation

This API requires the following crate features to be activated: Comment, Document

impl Document

pub fn create_element(&self, local_name: &str) -> Result<Element, JsValue>

The createElement() method

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn create_element_with_str(
&self,
local_name: &str,
options: &str
) -> Result<Element, JsValue>

The createElement() method

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn create_element_ns(
&self,
namespace: Option<&str>,
qualified_name: &str
) -> Result<Element, JsValue>

The createElementNS() method

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn create_element_ns_with_str(
&self,
namespace: Option<&str>,
qualified_name: &str,
options: &str
) -> Result<Element, JsValue>

The createElementNS() method

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn create_event(&self, interface: &str) -> Result<Event, JsValue>

The createEvent() method

MDN Documentation

This API requires the following crate features to be activated: Document, Event

impl Document

pub fn create_text_node(&self, data: &str) -> Text

The createTextNode() method

MDN Documentation

This API requires the following crate features to be activated: Document, Text

impl Document

pub fn enable_style_sheets_for_set(&self, name: Option<&str>)

The enableStyleSheetsForSet() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn exit_fullscreen(&self)

The exitFullscreen() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn exit_pointer_lock(&self)

The exitPointerLock() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn get_animations(&self) -> Array

The getAnimations() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn get_element_by_id(&self, element_id: &str) -> Option<Element>

The getElementById() method

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn get_elements_by_class_name(&self, class_names: &str) -> HtmlCollection

The getElementsByClassName() method

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn get_elements_by_name(&self, element_name: &str) -> NodeList

The getElementsByName() method

MDN Documentation

This API requires the following crate features to be activated: Document, NodeList

impl Document

pub fn get_elements_by_tag_name(&self, local_name: &str) -> HtmlCollection

The getElementsByTagName() method

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn get_elements_by_tag_name_ns(
&self,
namespace: Option<&str>,
local_name: &str
) -> Result<HtmlCollection, JsValue>

The getElementsByTagNameNS() method

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn has_focus(&self) -> Result<bool, JsValue>

The hasFocus() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn import_node(&self, node: &Node) -> Result<Node, JsValue>

The importNode() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn import_node_with_deep(
&self,
node: &Node,
deep: bool
) -> Result<Node, JsValue>

The importNode() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn query_selector(
&self,
selectors: &str
) -> Result<Option<Element>, JsValue>

The querySelector() method

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn query_selector_all(&self, selectors: &str) -> Result<NodeList, JsValue>

The querySelectorAll() method

MDN Documentation

This API requires the following crate features to be activated: Document, NodeList

impl Document

pub fn release_capture(&self)

The releaseCapture() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn url(&self) -> Result<String, JsValue>

The URL getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn document_uri(&self) -> Result<String, JsValue>

The documentURI getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn compat_mode(&self) -> String

The compatMode getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn character_set(&self) -> String

The characterSet getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn charset(&self) -> String

The charset getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn input_encoding(&self) -> String

The inputEncoding getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn content_type(&self) -> String

The contentType getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn document_element(&self) -> Option<Element>

The documentElement getter

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn referrer(&self) -> String

The referrer getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn last_modified(&self) -> String

The lastModified getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ready_state(&self) -> String

The readyState getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn title(&self) -> String

The title getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_title(&self, title: &str)

The title setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn dir(&self) -> String

The dir getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_dir(&self, dir: &str)

The dir setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn body(&self) -> Option<HtmlElement>

The body getter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlElement

impl Document

pub fn set_body(&self, body: Option<&HtmlElement>)

The body setter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlElement

impl Document

pub fn images(&self) -> HtmlCollection

The images getter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn embeds(&self) -> HtmlCollection

The embeds getter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn plugins(&self) -> HtmlCollection

The plugins getter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn links(&self) -> HtmlCollection

The links getter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn forms(&self) -> HtmlCollection

The forms getter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn scripts(&self) -> HtmlCollection

The scripts getter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn default_view(&self) -> Option<Window>

The defaultView getter

MDN Documentation

This API requires the following crate features to be activated: Document, Window

impl Document

pub fn onreadystatechange(&self) -> Option<Function>

The onreadystatechange getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onreadystatechange(&self, onreadystatechange: Option<&Function>)

The onreadystatechange setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onbeforescriptexecute(&self) -> Option<Function>

The onbeforescriptexecute getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onbeforescriptexecute(
&self,
onbeforescriptexecute: Option<&Function>
)

The onbeforescriptexecute setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onafterscriptexecute(&self) -> Option<Function>

The onafterscriptexecute getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onafterscriptexecute(&self, onafterscriptexecute: Option<&Function>)

The onafterscriptexecute setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onselectionchange(&self) -> Option<Function>

The onselectionchange getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onselectionchange(&self, onselectionchange: Option<&Function>)

The onselectionchange setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn current_script(&self) -> Option<Element>

The currentScript getter

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn anchors(&self) -> HtmlCollection

The anchors getter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn applets(&self) -> HtmlCollection

The applets getter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn fullscreen(&self) -> bool

The fullscreen getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn fullscreen_enabled(&self) -> bool

The fullscreenEnabled getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onfullscreenchange(&self) -> Option<Function>

The onfullscreenchange getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onfullscreenchange(&self, onfullscreenchange: Option<&Function>)

The onfullscreenchange setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onfullscreenerror(&self) -> Option<Function>

The onfullscreenerror getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onfullscreenerror(&self, onfullscreenerror: Option<&Function>)

The onfullscreenerror setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpointerlockchange(&self) -> Option<Function>

The onpointerlockchange getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpointerlockchange(&self, onpointerlockchange: Option<&Function>)

The onpointerlockchange setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpointerlockerror(&self) -> Option<Function>

The onpointerlockerror getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpointerlockerror(&self, onpointerlockerror: Option<&Function>)

The onpointerlockerror setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn hidden(&self) -> bool

The hidden getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onvisibilitychange(&self) -> Option<Function>

The onvisibilitychange getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onvisibilitychange(&self, onvisibilitychange: Option<&Function>)

The onvisibilitychange setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn selected_style_sheet_set(&self) -> Option<String>

The selectedStyleSheetSet getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_selected_style_sheet_set(
&self,
selected_style_sheet_set: Option<&str>
)

The selectedStyleSheetSet setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn last_style_sheet_set(&self) -> Option<String>

The lastStyleSheetSet getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn preferred_style_sheet_set(&self) -> Option<String>

The preferredStyleSheetSet getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn scrolling_element(&self) -> Option<Element>

The scrollingElement getter

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn oncopy(&self) -> Option<Function>

The oncopy getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_oncopy(&self, oncopy: Option<&Function>)

The oncopy setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn oncut(&self) -> Option<Function>

The oncut getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_oncut(&self, oncut: Option<&Function>)

The oncut setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpaste(&self) -> Option<Function>

The onpaste getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpaste(&self, onpaste: Option<&Function>)

The onpaste setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn element_from_point(&self, x: f32, y: f32) -> Option<Element>

The elementFromPoint() method

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn elements_from_point(&self, x: f32, y: f32) -> Array

The elementsFromPoint() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn active_element(&self) -> Option<Element>

The activeElement getter

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn pointer_lock_element(&self) -> Option<Element>

The pointerLockElement getter

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn fullscreen_element(&self) -> Option<Element>

The fullscreenElement getter

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn get_box_quads(&self) -> Result<Array, JsValue>

The getBoxQuads() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onabort(&self) -> Option<Function>

The onabort getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onabort(&self, onabort: Option<&Function>)

The onabort setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onblur(&self) -> Option<Function>

The onblur getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onblur(&self, onblur: Option<&Function>)

The onblur setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onfocus(&self) -> Option<Function>

The onfocus getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onfocus(&self, onfocus: Option<&Function>)

The onfocus setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onauxclick(&self) -> Option<Function>

The onauxclick getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onauxclick(&self, onauxclick: Option<&Function>)

The onauxclick setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn oncanplay(&self) -> Option<Function>

The oncanplay getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_oncanplay(&self, oncanplay: Option<&Function>)

The oncanplay setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn oncanplaythrough(&self) -> Option<Function>

The oncanplaythrough getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_oncanplaythrough(&self, oncanplaythrough: Option<&Function>)

The oncanplaythrough setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onchange(&self) -> Option<Function>

The onchange getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onchange(&self, onchange: Option<&Function>)

The onchange setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onclick(&self) -> Option<Function>

The onclick getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onclick(&self, onclick: Option<&Function>)

The onclick setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onclose(&self) -> Option<Function>

The onclose getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onclose(&self, onclose: Option<&Function>)

The onclose setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn oncontextmenu(&self) -> Option<Function>

The oncontextmenu getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_oncontextmenu(&self, oncontextmenu: Option<&Function>)

The oncontextmenu setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ondblclick(&self) -> Option<Function>

The ondblclick getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ondblclick(&self, ondblclick: Option<&Function>)

The ondblclick setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ondrag(&self) -> Option<Function>

The ondrag getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ondrag(&self, ondrag: Option<&Function>)

The ondrag setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ondragend(&self) -> Option<Function>

The ondragend getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ondragend(&self, ondragend: Option<&Function>)

The ondragend setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ondragenter(&self) -> Option<Function>

The ondragenter getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ondragenter(&self, ondragenter: Option<&Function>)

The ondragenter setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ondragexit(&self) -> Option<Function>

The ondragexit getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ondragexit(&self, ondragexit: Option<&Function>)

The ondragexit setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ondragleave(&self) -> Option<Function>

The ondragleave getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ondragleave(&self, ondragleave: Option<&Function>)

The ondragleave setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ondragover(&self) -> Option<Function>

The ondragover getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ondragover(&self, ondragover: Option<&Function>)

The ondragover setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ondragstart(&self) -> Option<Function>

The ondragstart getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ondragstart(&self, ondragstart: Option<&Function>)

The ondragstart setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ondrop(&self) -> Option<Function>

The ondrop getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ondrop(&self, ondrop: Option<&Function>)

The ondrop setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ondurationchange(&self) -> Option<Function>

The ondurationchange getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ondurationchange(&self, ondurationchange: Option<&Function>)

The ondurationchange setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onemptied(&self) -> Option<Function>

The onemptied getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onemptied(&self, onemptied: Option<&Function>)

The onemptied setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onended(&self) -> Option<Function>

The onended getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onended(&self, onended: Option<&Function>)

The onended setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn oninput(&self) -> Option<Function>

The oninput getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_oninput(&self, oninput: Option<&Function>)

The oninput setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn oninvalid(&self) -> Option<Function>

The oninvalid getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_oninvalid(&self, oninvalid: Option<&Function>)

The oninvalid setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onkeydown(&self) -> Option<Function>

The onkeydown getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onkeydown(&self, onkeydown: Option<&Function>)

The onkeydown setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onkeypress(&self) -> Option<Function>

The onkeypress getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onkeypress(&self, onkeypress: Option<&Function>)

The onkeypress setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onkeyup(&self) -> Option<Function>

The onkeyup getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onkeyup(&self, onkeyup: Option<&Function>)

The onkeyup setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onload(&self) -> Option<Function>

The onload getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onload(&self, onload: Option<&Function>)

The onload setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onloadeddata(&self) -> Option<Function>

The onloadeddata getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onloadeddata(&self, onloadeddata: Option<&Function>)

The onloadeddata setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onloadedmetadata(&self) -> Option<Function>

The onloadedmetadata getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onloadedmetadata(&self, onloadedmetadata: Option<&Function>)

The onloadedmetadata setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onloadend(&self) -> Option<Function>

The onloadend getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onloadend(&self, onloadend: Option<&Function>)

The onloadend setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onloadstart(&self) -> Option<Function>

The onloadstart getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onloadstart(&self, onloadstart: Option<&Function>)

The onloadstart setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onmousedown(&self) -> Option<Function>

The onmousedown getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onmousedown(&self, onmousedown: Option<&Function>)

The onmousedown setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onmouseenter(&self) -> Option<Function>

The onmouseenter getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onmouseenter(&self, onmouseenter: Option<&Function>)

The onmouseenter setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onmouseleave(&self) -> Option<Function>

The onmouseleave getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onmouseleave(&self, onmouseleave: Option<&Function>)

The onmouseleave setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onmousemove(&self) -> Option<Function>

The onmousemove getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onmousemove(&self, onmousemove: Option<&Function>)

The onmousemove setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onmouseout(&self) -> Option<Function>

The onmouseout getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onmouseout(&self, onmouseout: Option<&Function>)

The onmouseout setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onmouseover(&self) -> Option<Function>

The onmouseover getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onmouseover(&self, onmouseover: Option<&Function>)

The onmouseover setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onmouseup(&self) -> Option<Function>

The onmouseup getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onmouseup(&self, onmouseup: Option<&Function>)

The onmouseup setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onwheel(&self) -> Option<Function>

The onwheel getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onwheel(&self, onwheel: Option<&Function>)

The onwheel setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpause(&self) -> Option<Function>

The onpause getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpause(&self, onpause: Option<&Function>)

The onpause setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onplay(&self) -> Option<Function>

The onplay getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onplay(&self, onplay: Option<&Function>)

The onplay setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onplaying(&self) -> Option<Function>

The onplaying getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onplaying(&self, onplaying: Option<&Function>)

The onplaying setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onprogress(&self) -> Option<Function>

The onprogress getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onprogress(&self, onprogress: Option<&Function>)

The onprogress setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onratechange(&self) -> Option<Function>

The onratechange getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onratechange(&self, onratechange: Option<&Function>)

The onratechange setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onreset(&self) -> Option<Function>

The onreset getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onreset(&self, onreset: Option<&Function>)

The onreset setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onresize(&self) -> Option<Function>

The onresize getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onresize(&self, onresize: Option<&Function>)

The onresize setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onscroll(&self) -> Option<Function>

The onscroll getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onscroll(&self, onscroll: Option<&Function>)

The onscroll setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onseeked(&self) -> Option<Function>

The onseeked getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onseeked(&self, onseeked: Option<&Function>)

The onseeked setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onseeking(&self) -> Option<Function>

The onseeking getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onseeking(&self, onseeking: Option<&Function>)

The onseeking setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onselect(&self) -> Option<Function>

The onselect getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onselect(&self, onselect: Option<&Function>)

The onselect setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onshow(&self) -> Option<Function>

The onshow getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onshow(&self, onshow: Option<&Function>)

The onshow setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onstalled(&self) -> Option<Function>

The onstalled getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onstalled(&self, onstalled: Option<&Function>)

The onstalled setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onsubmit(&self) -> Option<Function>

The onsubmit getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onsubmit(&self, onsubmit: Option<&Function>)

The onsubmit setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onsuspend(&self) -> Option<Function>

The onsuspend getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onsuspend(&self, onsuspend: Option<&Function>)

The onsuspend setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ontimeupdate(&self) -> Option<Function>

The ontimeupdate getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ontimeupdate(&self, ontimeupdate: Option<&Function>)

The ontimeupdate setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onvolumechange(&self) -> Option<Function>

The onvolumechange getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onvolumechange(&self, onvolumechange: Option<&Function>)

The onvolumechange setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onwaiting(&self) -> Option<Function>

The onwaiting getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onwaiting(&self, onwaiting: Option<&Function>)

The onwaiting setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onselectstart(&self) -> Option<Function>

The onselectstart getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onselectstart(&self, onselectstart: Option<&Function>)

The onselectstart setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ontoggle(&self) -> Option<Function>

The ontoggle getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ontoggle(&self, ontoggle: Option<&Function>)

The ontoggle setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpointercancel(&self) -> Option<Function>

The onpointercancel getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpointercancel(&self, onpointercancel: Option<&Function>)

The onpointercancel setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpointerdown(&self) -> Option<Function>

The onpointerdown getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpointerdown(&self, onpointerdown: Option<&Function>)

The onpointerdown setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpointerup(&self) -> Option<Function>

The onpointerup getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpointerup(&self, onpointerup: Option<&Function>)

The onpointerup setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpointermove(&self) -> Option<Function>

The onpointermove getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpointermove(&self, onpointermove: Option<&Function>)

The onpointermove setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpointerout(&self) -> Option<Function>

The onpointerout getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpointerout(&self, onpointerout: Option<&Function>)

The onpointerout setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpointerover(&self) -> Option<Function>

The onpointerover getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpointerover(&self, onpointerover: Option<&Function>)

The onpointerover setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpointerenter(&self) -> Option<Function>

The onpointerenter getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpointerenter(&self, onpointerenter: Option<&Function>)

The onpointerenter setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onpointerleave(&self) -> Option<Function>

The onpointerleave getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onpointerleave(&self, onpointerleave: Option<&Function>)

The onpointerleave setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ongotpointercapture(&self) -> Option<Function>

The ongotpointercapture getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ongotpointercapture(&self, ongotpointercapture: Option<&Function>)

The ongotpointercapture setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onlostpointercapture(&self) -> Option<Function>

The onlostpointercapture getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onlostpointercapture(&self, onlostpointercapture: Option<&Function>)

The onlostpointercapture setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onanimationcancel(&self) -> Option<Function>

The onanimationcancel getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onanimationcancel(&self, onanimationcancel: Option<&Function>)

The onanimationcancel setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onanimationend(&self) -> Option<Function>

The onanimationend getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onanimationend(&self, onanimationend: Option<&Function>)

The onanimationend setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onanimationiteration(&self) -> Option<Function>

The onanimationiteration getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onanimationiteration(&self, onanimationiteration: Option<&Function>)

The onanimationiteration setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onanimationstart(&self) -> Option<Function>

The onanimationstart getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onanimationstart(&self, onanimationstart: Option<&Function>)

The onanimationstart setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ontransitioncancel(&self) -> Option<Function>

The ontransitioncancel getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ontransitioncancel(&self, ontransitioncancel: Option<&Function>)

The ontransitioncancel setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ontransitionend(&self) -> Option<Function>

The ontransitionend getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ontransitionend(&self, ontransitionend: Option<&Function>)

The ontransitionend setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ontransitionrun(&self) -> Option<Function>

The ontransitionrun getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ontransitionrun(&self, ontransitionrun: Option<&Function>)

The ontransitionrun setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ontransitionstart(&self) -> Option<Function>

The ontransitionstart getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ontransitionstart(&self, ontransitionstart: Option<&Function>)

The ontransitionstart setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onwebkitanimationend(&self) -> Option<Function>

The onwebkitanimationend getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onwebkitanimationend(&self, onwebkitanimationend: Option<&Function>)

The onwebkitanimationend setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onwebkitanimationiteration(&self) -> Option<Function>

The onwebkitanimationiteration getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onwebkitanimationiteration(
&self,
onwebkitanimationiteration: Option<&Function>
)

The onwebkitanimationiteration setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onwebkitanimationstart(&self) -> Option<Function>

The onwebkitanimationstart getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onwebkitanimationstart(
&self,
onwebkitanimationstart: Option<&Function>
)

The onwebkitanimationstart setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onwebkittransitionend(&self) -> Option<Function>

The onwebkittransitionend getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onwebkittransitionend(
&self,
onwebkittransitionend: Option<&Function>
)

The onwebkittransitionend setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_node_0(&self) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn append_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn append_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn append_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn append_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn append_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn append_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn append_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_str_0(&self) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn append_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_node_0(&self) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn prepend_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn prepend_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn prepend_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn prepend_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn prepend_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn prepend_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Document

pub fn prepend_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_str_0(&self) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn prepend_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn children(&self) -> HtmlCollection

The children getter

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlCollection

impl Document

pub fn first_element_child(&self) -> Option<Element>

The firstElementChild getter

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn last_element_child(&self) -> Option<Element>

The lastElementChild getter

MDN Documentation

This API requires the following crate features to be activated: Document, Element

impl Document

pub fn child_element_count(&self) -> u32

The childElementCount getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ontouchstart(&self) -> Option<Function>

The ontouchstart getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ontouchstart(&self, ontouchstart: Option<&Function>)

The ontouchstart setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ontouchend(&self) -> Option<Function>

The ontouchend getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ontouchend(&self, ontouchend: Option<&Function>)

The ontouchend setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ontouchmove(&self) -> Option<Function>

The ontouchmove getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ontouchmove(&self, ontouchmove: Option<&Function>)

The ontouchmove setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn ontouchcancel(&self) -> Option<Function>

The ontouchcancel getter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn set_ontouchcancel(&self, ontouchcancel: Option<&Function>)

The ontouchcancel setter

MDN Documentation

This API requires the following crate features to be activated: Document

impl Document

pub fn create_ns_resolver(&self, node_resolver: &Node) -> Node

The createNSResolver() method

MDN Documentation

This API requires the following crate features to be activated: Document, Node