Struct virtual_dom_rs::DomTokenList

#[repr(transparent)]

pub struct DomTokenList { /* fields omitted */ }

The DOMTokenList object

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

Methods

impl DomTokenList

pub fn add(&self, tokens: &Array) -> Result<(), JsValue>

The add() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn add_0(&self) -> Result<(), JsValue>

The add() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn add_1(&self, tokens_1: &str) -> Result<(), JsValue>

The add() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn add_2(&self, tokens_1: &str, tokens_2: &str) -> Result<(), JsValue>

The add() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn add_3(
&self,
tokens_1: &str,
tokens_2: &str,
tokens_3: &str
) -> Result<(), JsValue>

The add() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn add_4(
&self,
tokens_1: &str,
tokens_2: &str,
tokens_3: &str,
tokens_4: &str
) -> Result<(), JsValue>

The add() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn add_5(
&self,
tokens_1: &str,
tokens_2: &str,
tokens_3: &str,
tokens_4: &str,
tokens_5: &str
) -> Result<(), JsValue>

The add() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn add_6(
&self,
tokens_1: &str,
tokens_2: &str,
tokens_3: &str,
tokens_4: &str,
tokens_5: &str,
tokens_6: &str
) -> Result<(), JsValue>

The add() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn add_7(
&self,
tokens_1: &str,
tokens_2: &str,
tokens_3: &str,
tokens_4: &str,
tokens_5: &str,
tokens_6: &str,
tokens_7: &str
) -> Result<(), JsValue>

The add() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn contains(&self, token: &str) -> bool

The contains() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn item(&self, index: u32) -> Option<String>

The item() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn remove(&self, tokens: &Array) -> Result<(), JsValue>

The remove() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn remove_0(&self) -> Result<(), JsValue>

The remove() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn remove_1(&self, tokens_1: &str) -> Result<(), JsValue>

The remove() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn remove_2(&self, tokens_1: &str, tokens_2: &str) -> Result<(), JsValue>

The remove() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn remove_3(
&self,
tokens_1: &str,
tokens_2: &str,
tokens_3: &str
) -> Result<(), JsValue>

The remove() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn remove_4(
&self,
tokens_1: &str,
tokens_2: &str,
tokens_3: &str,
tokens_4: &str
) -> Result<(), JsValue>

The remove() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn remove_5(
&self,
tokens_1: &str,
tokens_2: &str,
tokens_3: &str,
tokens_4: &str,
tokens_5: &str
) -> Result<(), JsValue>

The remove() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn remove_6(
&self,
tokens_1: &str,
tokens_2: &str,
tokens_3: &str,
tokens_4: &str,
tokens_5: &str,
tokens_6: &str
) -> Result<(), JsValue>

The remove() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn remove_7(
&self,
tokens_1: &str,
tokens_2: &str,
tokens_3: &str,
tokens_4: &str,
tokens_5: &str,
tokens_6: &str,
tokens_7: &str
) -> Result<(), JsValue>

The remove() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn replace(&self, token: &str, new_token: &str) -> Result<bool, JsValue>

The replace() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn supports(&self, token: &str) -> Result<bool, JsValue>

The supports() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn toggle(&self, token: &str) -> Result<bool, JsValue>

The toggle() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn toggle_with_force(
&self,
token: &str,
force: bool
) -> Result<bool, JsValue>

The toggle() method

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn get(&self, index: u32) -> Option<String>

The indexing getter

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn length(&self) -> u32

The length getter

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn value(&self) -> String

The value getter

MDN Documentation

This API requires the following crate features to be activated: DomTokenList

impl DomTokenList

pub fn set_value(&self, value: &str)

The value setter

MDN Documentation

This API requires the following crate features to be activated: DomTokenList