Struct virtual_dom_rs::Element

#[repr(transparent)]

pub struct Element { /* fields omitted */ }

The Element object

MDN Documentation

This API requires the following crate features to be activated: Element

Methods

impl Element

pub fn closest(&self, selector: &str) -> Result<Option<Element>, JsValue>

The closest() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn get_attribute(&self, name: &str) -> Option<String>

The getAttribute() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn get_attribute_ns(
&self,
namespace: Option<&str>,
local_name: &str
) -> Option<String>

The getAttributeNS() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn get_attribute_names(&self) -> Array

The getAttributeNames() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn has_attribute(&self, name: &str) -> bool

The hasAttribute() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn has_attribute_ns(
&self,
namespace: Option<&str>,
local_name: &str
) -> bool

The hasAttributeNS() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn has_attributes(&self) -> bool

The hasAttributes() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn has_pointer_capture(&self, pointer_id: i32) -> bool

The hasPointerCapture() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn insert_adjacent_element(
&self,
where_: &str,
element: &Element
) -> Result<Option<Element>, JsValue>

The insertAdjacentElement() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn insert_adjacent_html(
&self,
position: &str,
text: &str
) -> Result<(), JsValue>

The insertAdjacentHTML() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn insert_adjacent_text(
&self,
where_: &str,
data: &str
) -> Result<(), JsValue>

The insertAdjacentText() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn matches(&self, selector: &str) -> Result<bool, JsValue>

The matches() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn query_selector(
&self,
selectors: &str
) -> Result<Option<Element>, JsValue>

The querySelector() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn query_selector_all(&self, selectors: &str) -> Result<NodeList, JsValue>

The querySelectorAll() method

MDN Documentation

This API requires the following crate features to be activated: Element, NodeList

impl Element

pub fn release_capture(&self)

The releaseCapture() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn release_pointer_capture(&self, pointer_id: i32) -> Result<(), JsValue>

The releasePointerCapture() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn remove_attribute(&self, name: &str) -> Result<(), JsValue>

The removeAttribute() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn remove_attribute_ns(
&self,
namespace: Option<&str>,
local_name: &str
) -> Result<(), JsValue>

The removeAttributeNS() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn request_fullscreen(&self) -> Result<(), JsValue>

The requestFullscreen() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn request_pointer_lock(&self)

The requestPointerLock() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_with_x_and_y(&self, x: f64, y: f64)

The scroll() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll(&self)

The scroll() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_by_with_x_and_y(&self, x: f64, y: f64)

The scrollBy() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_by(&self)

The scrollBy() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_into_view(&self)

The scrollIntoView() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_into_view_with_bool(&self, arg: bool)

The scrollIntoView() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_to_with_x_and_y(&self, x: f64, y: f64)

The scrollTo() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_to(&self)

The scrollTo() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_attribute(&self, name: &str, value: &str) -> Result<(), JsValue>

The setAttribute() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_attribute_ns(
&self,
namespace: Option<&str>,
name: &str,
value: &str
) -> Result<(), JsValue>

The setAttributeNS() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_capture(&self)

The setCapture() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_capture_with_retarget_to_element(&self, retarget_to_element: bool)

The setCapture() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_pointer_capture(&self, pointer_id: i32) -> Result<(), JsValue>

The setPointerCapture() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn toggle_attribute(&self, name: &str) -> Result<bool, JsValue>

The toggleAttribute() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn toggle_attribute_with_force(
&self,
name: &str,
force: bool
) -> Result<bool, JsValue>

The toggleAttribute() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn webkit_matches_selector(&self, selector: &str) -> Result<bool, JsValue>

The webkitMatchesSelector() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn namespace_uri(&self) -> Option<String>

The namespaceURI getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prefix(&self) -> Option<String>

The prefix getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn local_name(&self) -> String

The localName getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn tag_name(&self) -> String

The tagName getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn id(&self) -> String

The id getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_id(&self, id: &str)

The id setter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn class_name(&self) -> String

The className getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_class_name(&self, class_name: &str)

The className setter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn class_list(&self) -> DomTokenList

The classList getter

MDN Documentation

This API requires the following crate features to be activated: DomTokenList, Element

impl Element

pub fn scroll_top(&self) -> i32

The scrollTop getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_scroll_top(&self, scroll_top: i32)

The scrollTop setter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_left(&self) -> i32

The scrollLeft getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_scroll_left(&self, scroll_left: i32)

The scrollLeft setter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_width(&self) -> i32

The scrollWidth getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_height(&self) -> i32

The scrollHeight getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn client_top(&self) -> i32

The clientTop getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn client_left(&self) -> i32

The clientLeft getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn client_width(&self) -> i32

The clientWidth getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn client_height(&self) -> i32

The clientHeight getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn inner_html(&self) -> String

The innerHTML getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_inner_html(&self, inner_html: &str)

The innerHTML setter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn outer_html(&self) -> String

The outerHTML getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_outer_html(&self, outer_html: &str)

The outerHTML setter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn slot(&self) -> String

The slot getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_slot(&self, slot: &str)

The slot setter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node_0(&self) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn after_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn after_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn after_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn after_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn after_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn after_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn after_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_0(&self) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node_0(&self) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn before_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn before_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn before_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn before_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn before_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn before_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn before_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_0(&self) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn remove(&self)

The remove() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node_0(&self) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn replace_with_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn replace_with_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn replace_with_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn replace_with_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn replace_with_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn replace_with_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn replace_with_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_0(&self) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn get_box_quads(&self) -> Result<Array, JsValue>

The getBoxQuads() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn previous_element_sibling(&self) -> Option<Element>

The previousElementSibling getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn next_element_sibling(&self) -> Option<Element>

The nextElementSibling getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node_0(&self) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn append_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn append_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn append_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn append_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn append_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn append_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn append_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_0(&self) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The append() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node_0(&self) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn prepend_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn prepend_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn prepend_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn prepend_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn prepend_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn prepend_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Element

pub fn prepend_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_0(&self) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The prepend() method

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn children(&self) -> HtmlCollection

The children getter

MDN Documentation

This API requires the following crate features to be activated: Element, HtmlCollection

impl Element

pub fn first_element_child(&self) -> Option<Element>

The firstElementChild getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn last_element_child(&self) -> Option<Element>

The lastElementChild getter

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn child_element_count(&self) -> u32

The childElementCount getter

MDN Documentation

This API requires the following crate features to be activated: Element