Struct virtual_dom_rs::HtmlCollection

#[repr(transparent)]
pub struct HtmlCollection { /* fields omitted */ }

The HTMLCollection object

MDN Documentation

This API requires the following crate features to be activated: HtmlCollection

Methods

impl HtmlCollection

pub fn item(&self, index: u32) -> Option<Element>

The item() method

MDN Documentation

This API requires the following crate features to be activated: Element, HtmlCollection

impl HtmlCollection

pub fn named_item(&self, name: &str) -> Option<Element>

The namedItem() method

MDN Documentation

This API requires the following crate features to be activated: Element, HtmlCollection

impl HtmlCollection

pub fn get_with_index(&self, index: u32) -> Option<Element>

The indexing getter

This API requires the following crate features to be activated: Element, HtmlCollection

impl HtmlCollection

pub fn get_with_name(&self, name: &str) -> Option<Element>

The indexing getter

This API requires the following crate features to be activated: Element, HtmlCollection

impl HtmlCollection

pub fn length(&self) -> u32

The length getter

MDN Documentation

This API requires the following crate features to be activated: HtmlCollection