Struct virtual_dom_rs::HtmlInputElement

#[repr(transparent)]

pub struct HtmlInputElement { /* fields omitted */ }

The HTMLInputElement object

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

Methods

impl HtmlInputElement

pub fn check_validity(&self) -> bool

The checkValidity() method

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn report_validity(&self) -> bool

The reportValidity() method

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn select(&self)

The select() method

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_custom_validity(&self, error: &str)

The setCustomValidity() method

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_range_text(&self, replacement: &str) -> Result<(), JsValue>

The setRangeText() method

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_range_text_with_start_and_end(
&self,
replacement: &str,
start: u32,
end: u32
) -> Result<(), JsValue>

The setRangeText() method

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_selection_range(&self, start: u32, end: u32) -> Result<(), JsValue>

The setSelectionRange() method

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_selection_range_with_direction(
&self,
start: u32,
end: u32,
direction: &str
) -> Result<(), JsValue>

The setSelectionRange() method

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn accept(&self) -> String

The accept getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_accept(&self, accept: &str)

The accept setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn alt(&self) -> String

The alt getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_alt(&self, alt: &str)

The alt setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn autocomplete(&self) -> String

The autocomplete getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_autocomplete(&self, autocomplete: &str)

The autocomplete setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn autofocus(&self) -> bool

The autofocus getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_autofocus(&self, autofocus: bool)

The autofocus setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn default_checked(&self) -> bool

The defaultChecked getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_default_checked(&self, default_checked: bool)

The defaultChecked setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn checked(&self) -> bool

The checked getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_checked(&self, checked: bool)

The checked setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn disabled(&self) -> bool

The disabled getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_disabled(&self, disabled: bool)

The disabled setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn form_action(&self) -> String

The formAction getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_form_action(&self, form_action: &str)

The formAction setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn form_enctype(&self) -> String

The formEnctype getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_form_enctype(&self, form_enctype: &str)

The formEnctype setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn form_method(&self) -> String

The formMethod getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_form_method(&self, form_method: &str)

The formMethod setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn form_no_validate(&self) -> bool

The formNoValidate getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_form_no_validate(&self, form_no_validate: bool)

The formNoValidate setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn form_target(&self) -> String

The formTarget getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_form_target(&self, form_target: &str)

The formTarget setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn height(&self) -> u32

The height getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_height(&self, height: u32)

The height setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn indeterminate(&self) -> bool

The indeterminate getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_indeterminate(&self, indeterminate: bool)

The indeterminate setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn input_mode(&self) -> String

The inputMode getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_input_mode(&self, input_mode: &str)

The inputMode setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn list(&self) -> Option<HtmlElement>

The list getter

MDN Documentation

This API requires the following crate features to be activated: HtmlElement, HtmlInputElement

impl HtmlInputElement

pub fn max(&self) -> String

The max getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_max(&self, max: &str)

The max setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn max_length(&self) -> i32

The maxLength getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_max_length(&self, max_length: i32)

The maxLength setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn min(&self) -> String

The min getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_min(&self, min: &str)

The min setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn min_length(&self) -> i32

The minLength getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_min_length(&self, min_length: i32)

The minLength setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn multiple(&self) -> bool

The multiple getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_multiple(&self, multiple: bool)

The multiple setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn name(&self) -> String

The name getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_name(&self, name: &str)

The name setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn pattern(&self) -> String

The pattern getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_pattern(&self, pattern: &str)

The pattern setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn placeholder(&self) -> String

The placeholder getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_placeholder(&self, placeholder: &str)

The placeholder setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn read_only(&self) -> bool

The readOnly getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_read_only(&self, read_only: bool)

The readOnly setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn required(&self) -> bool

The required getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_required(&self, required: bool)

The required setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn size(&self) -> u32

The size getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_size(&self, size: u32)

The size setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn src(&self) -> String

The src getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_src(&self, src: &str)

The src setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn step(&self) -> String

The step getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_step(&self, step: &str)

The step setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn type_(&self) -> String

The type getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_type(&self, type_: &str)

The type setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn default_value(&self) -> String

The defaultValue getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_default_value(&self, default_value: &str)

The defaultValue setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn value(&self) -> String

The value getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_value(&self, value: &str)

The value setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn value_as_number(&self) -> f64

The valueAsNumber getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_value_as_number(&self, value_as_number: f64)

The valueAsNumber setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn width(&self) -> u32

The width getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_width(&self, width: u32)

The width setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn will_validate(&self) -> bool

The willValidate getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn validation_message(&self) -> Result<String, JsValue>

The validationMessage getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn labels(&self) -> Option<NodeList>

The labels getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement, NodeList

impl HtmlInputElement

pub fn selection_direction(&self) -> Result<Option<String>, JsValue>

The selectionDirection getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_selection_direction(
&self,
selection_direction: Option<&str>
) -> Result<(), JsValue>

The selectionDirection setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn align(&self) -> String

The align getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_align(&self, align: &str)

The align setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn use_map(&self) -> String

The useMap getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_use_map(&self, use_map: &str)

The useMap setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn webkit_entries(&self) -> Array

The webkitEntries getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn webkitdirectory(&self) -> bool

The webkitdirectory getter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_webkitdirectory(&self, webkitdirectory: bool)

The webkitdirectory setter

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement