Struct virtual_dom_rs::MouseEvent

#[repr(transparent)]

pub struct MouseEvent { /* fields omitted */ }

The MouseEvent object

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

Methods

impl MouseEvent

pub fn new(type_arg: &str) -> Result<MouseEvent, JsValue>

The new MouseEvent(..) constructor, creating a new instance of MouseEvent

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn get_modifier_state(&self, key_arg: &str) -> bool

The getModifierState() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn init_mouse_event(&self, type_arg: &str)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg(
&self,
type_arg: &str,
can_bubble_arg: bool
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg_and_alt_key_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool,
alt_key_arg: bool
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg_and_alt_key_arg_and_shift_key_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool,
alt_key_arg: bool,
shift_key_arg: bool
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg_and_alt_key_arg_and_shift_key_arg_and_meta_key_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool,
alt_key_arg: bool,
shift_key_arg: bool,
meta_key_arg: bool
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg_and_alt_key_arg_and_shift_key_arg_and_meta_key_arg_and_button_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool,
alt_key_arg: bool,
shift_key_arg: bool,
meta_key_arg: bool,
button_arg: i16
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg_and_alt_key_arg_and_shift_key_arg_and_meta_key_arg_and_button_arg_and_related_target_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool,
alt_key_arg: bool,
shift_key_arg: bool,
meta_key_arg: bool,
button_arg: i16,
related_target_arg: Option<&EventTarget>
)

The initMouseEvent() method

MDN Documentation

This API requires the following crate features to be activated: EventTarget, MouseEvent, Window

impl MouseEvent

pub fn screen_x(&self) -> i32

The screenX getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn screen_y(&self) -> i32

The screenY getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn client_x(&self) -> i32

The clientX getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn client_y(&self) -> i32

The clientY getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn x(&self) -> i32

The x getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn y(&self) -> i32

The y getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn offset_x(&self) -> i32

The offsetX getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn offset_y(&self) -> i32

The offsetY getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn ctrl_key(&self) -> bool

The ctrlKey getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn shift_key(&self) -> bool

The shiftKey getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn alt_key(&self) -> bool

The altKey getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn meta_key(&self) -> bool

The metaKey getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn button(&self) -> i16

The button getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn buttons(&self) -> u16

The buttons getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn related_target(&self) -> Option<EventTarget>

The relatedTarget getter

MDN Documentation

This API requires the following crate features to be activated: EventTarget, MouseEvent

impl MouseEvent

pub fn region(&self) -> Option<String>

The region getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn movement_x(&self) -> i32

The movementX getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn movement_y(&self) -> i32

The movementY getter

MDN Documentation

This API requires the following crate features to be activated: MouseEvent