Struct virtual_dom_rs::Node

#[repr(transparent)]

pub struct Node { /* fields omitted */ }

The Node object

MDN Documentation

This API requires the following crate features to be activated: Node

Methods

impl Node

pub fn append_child(&self, node: &Node) -> Result<Node, JsValue>

The appendChild() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn clone_node(&self) -> Result<Node, JsValue>

The cloneNode() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn clone_node_with_deep(&self, deep: bool) -> Result<Node, JsValue>

The cloneNode() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn compare_document_position(&self, other: &Node) -> u16

The compareDocumentPosition() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn contains(&self, other: Option<&Node>) -> bool

The contains() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn get_root_node(&self) -> Node

The getRootNode() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn has_child_nodes(&self) -> bool

The hasChildNodes() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn insert_before(
&self,
node: &Node,
child: Option<&Node>
) -> Result<Node, JsValue>

The insertBefore() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn is_default_namespace(&self, namespace: Option<&str>) -> bool

The isDefaultNamespace() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn is_equal_node(&self, node: Option<&Node>) -> bool

The isEqualNode() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn is_same_node(&self, node: Option<&Node>) -> bool

The isSameNode() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn lookup_namespace_uri(&self, prefix: Option<&str>) -> Option<String>

The lookupNamespaceURI() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn lookup_prefix(&self, namespace: Option<&str>) -> Option<String>

The lookupPrefix() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn normalize(&self)

The normalize() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn remove_child(&self, child: &Node) -> Result<Node, JsValue>

The removeChild() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn replace_child(&self, node: &Node, child: &Node) -> Result<Node, JsValue>

The replaceChild() method

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn node_type(&self) -> u16

The nodeType getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn node_name(&self) -> String

The nodeName getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn base_uri(&self) -> Result<Option<String>, JsValue>

The baseURI getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn is_connected(&self) -> bool

The isConnected getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn owner_document(&self) -> Option<Document>

The ownerDocument getter

MDN Documentation

This API requires the following crate features to be activated: Document, Node

impl Node

pub fn parent_node(&self) -> Option<Node>

The parentNode getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn parent_element(&self) -> Option<Element>

The parentElement getter

MDN Documentation

This API requires the following crate features to be activated: Element, Node

impl Node

pub fn child_nodes(&self) -> NodeList

The childNodes getter

MDN Documentation

This API requires the following crate features to be activated: Node, NodeList

impl Node

pub fn first_child(&self) -> Option<Node>

The firstChild getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn last_child(&self) -> Option<Node>

The lastChild getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn previous_sibling(&self) -> Option<Node>

The previousSibling getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn next_sibling(&self) -> Option<Node>

The nextSibling getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn node_value(&self) -> Option<String>

The nodeValue getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn set_node_value(&self, node_value: Option<&str>)

The nodeValue setter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn text_content(&self) -> Option<String>

The textContent getter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub fn set_text_content(&self, text_content: Option<&str>)

The textContent setter

MDN Documentation

This API requires the following crate features to be activated: Node

impl Node

pub const ELEMENT_NODE: u16

impl Node

pub const ATTRIBUTE_NODE: u16

impl Node

pub const TEXT_NODE: u16

impl Node

pub const CDATA_SECTION_NODE: u16

impl Node

pub const ENTITY_REFERENCE_NODE: u16

impl Node

pub const ENTITY_NODE: u16

impl Node

pub const PROCESSING_INSTRUCTION_NODE: u16

impl Node

pub const COMMENT_NODE: u16

impl Node

pub const DOCUMENT_NODE: u16

impl Node

pub const DOCUMENT_TYPE_NODE: u16

impl Node

pub const DOCUMENT_FRAGMENT_NODE: u16

impl Node

pub const NOTATION_NODE: u16

impl Node

pub const DOCUMENT_POSITION_DISCONNECTED: u16

impl Node

pub const DOCUMENT_POSITION_PRECEDING: u16

impl Node

pub const DOCUMENT_POSITION_FOLLOWING: u16

impl Node

pub const DOCUMENT_POSITION_CONTAINS: u16

impl Node

pub const DOCUMENT_POSITION_CONTAINED_BY: u16

impl Node

pub const DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC: u16