Struct virtual_dom_rs::NodeList

#[repr(transparent)]

pub struct NodeList { /* fields omitted */ }

The NodeList object

MDN Documentation

This API requires the following crate features to be activated: NodeList

Methods

impl NodeList

pub fn item(&self, index: u32) -> Option<Node>

The item() method

MDN Documentation

This API requires the following crate features to be activated: Node, NodeList

impl NodeList

pub fn get(&self, index: u32) -> Option<Node>

The indexing getter

This API requires the following crate features to be activated: Node, NodeList

impl NodeList

pub fn length(&self) -> u32

The length getter

MDN Documentation

This API requires the following crate features to be activated: NodeList