Struct virtual_dom_rs::Text

#[repr(transparent)]pub struct Text { /* fields omitted */ }

The Text class.

MDN Documentation

This API requires the following crate features to be activated: Text

Implementations

impl Text

pub fn whole_text(&self) -> Result<String, JsValue>

Getter for the wholeText field of this object.

MDN Documentation

This API requires the following crate features to be activated: Text

impl Text

pub fn new() -> Result<Text, JsValue>

The new Text(..) constructor, creating a new instance of Text.

MDN Documentation

This API requires the following crate features to be activated: Text

impl Text

pub fn new_with_data(data: &str) -> Result<Text, JsValue>

The new Text(..) constructor, creating a new instance of Text.

MDN Documentation

This API requires the following crate features to be activated: Text

impl Text

pub fn split_text(&self, offset: u32) -> Result<Text, JsValue>

The splitText() method.

MDN Documentation

This API requires the following crate features to be activated: Text

impl Text

pub fn get_box_quads(&self) -> Result<Array, JsValue>

The getBoxQuads() method.

MDN Documentation

This API requires the following crate features to be activated: Text