Struct virtual_dom_rs::Worker

#[repr(transparent)]
pub struct Worker { /* fields omitted */ }

The Worker object

MDN Documentation

This API requires the following crate features to be activated: Worker

Methods

impl Worker

pub fn new(script_url: &str) -> Result<Worker, JsValue>

The new Worker(..) constructor, creating a new instance of Worker

MDN Documentation

This API requires the following crate features to be activated: Worker

impl Worker

pub fn post_message(&self, message: &JsValue) -> Result<(), JsValue>

The postMessage() method

MDN Documentation

This API requires the following crate features to be activated: Worker

impl Worker

pub fn post_message_with_transfer(
&self,
message: &JsValue,
transfer: &JsValue
) -> Result<(), JsValue>

The postMessage() method

MDN Documentation

This API requires the following crate features to be activated: Worker

impl Worker

pub fn terminate(&self)

The terminate() method

MDN Documentation

This API requires the following crate features to be activated: Worker

impl Worker

pub fn onmessage(&self) -> Option<Function>

The onmessage getter

MDN Documentation

This API requires the following crate features to be activated: Worker

impl Worker

pub fn set_onmessage(&self, onmessage: Option<&Function>)

The onmessage setter

MDN Documentation

This API requires the following crate features to be activated: Worker

impl Worker

pub fn onmessageerror(&self) -> Option<Function>

The onmessageerror getter

MDN Documentation

This API requires the following crate features to be activated: Worker

impl Worker

pub fn set_onmessageerror(&self, onmessageerror: Option<&Function>)

The onmessageerror setter

MDN Documentation

This API requires the following crate features to be activated: Worker

impl Worker

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This API requires the following crate features to be activated: Worker

impl Worker

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This API requires the following crate features to be activated: Worker