Struct virtual_dom_rs::HtmlElement

#[repr(transparent)]
pub struct HtmlElement { /* fields omitted */ }

The HtmlElement class.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

Methods

impl HtmlElement

pub fn title(&self) -> String

Getter for the title field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_title(&self, value: &str)

Setter for the title field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn lang(&self) -> String

Getter for the lang field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_lang(&self, value: &str)

Setter for the lang field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn dir(&self) -> String

Getter for the dir field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_dir(&self, value: &str)

Setter for the dir field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn inner_text(&self) -> String

Getter for the innerText field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_inner_text(&self, value: &str)

Setter for the innerText field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn hidden(&self) -> bool

Getter for the hidden field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_hidden(&self, value: bool)

Setter for the hidden field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn tab_index(&self) -> i32

Getter for the tabIndex field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_tab_index(&self, value: i32)

Setter for the tabIndex field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn access_key(&self) -> String

Getter for the accessKey field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_access_key(&self, value: &str)

Setter for the accessKey field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn access_key_label(&self) -> String

Getter for the accessKeyLabel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn draggable(&self) -> bool

Getter for the draggable field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_draggable(&self, value: bool)

Setter for the draggable field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn content_editable(&self) -> String

Getter for the contentEditable field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_content_editable(&self, value: &str)

Setter for the contentEditable field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn is_content_editable(&self) -> bool

Getter for the isContentEditable field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn spellcheck(&self) -> bool

Getter for the spellcheck field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_spellcheck(&self, value: bool)

Setter for the spellcheck field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn offset_parent(&self) -> Option<Element>

Getter for the offsetParent field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn offset_top(&self) -> i32

Getter for the offsetTop field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn offset_left(&self) -> i32

Getter for the offsetLeft field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn offset_width(&self) -> i32

Getter for the offsetWidth field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn offset_height(&self) -> i32

Getter for the offsetHeight field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn oncopy(&self) -> Option<Function>

Getter for the oncopy field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_oncopy(&self, value: Option<&Function>)

Setter for the oncopy field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn oncut(&self) -> Option<Function>

Getter for the oncut field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_oncut(&self, value: Option<&Function>)

Setter for the oncut field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onpaste(&self) -> Option<Function>

Getter for the onpaste field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onpaste(&self, value: Option<&Function>)

Setter for the onpaste field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onabort(&self) -> Option<Function>

Getter for the onabort field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onabort(&self, value: Option<&Function>)

Setter for the onabort field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onblur(&self) -> Option<Function>

Getter for the onblur field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onblur(&self, value: Option<&Function>)

Setter for the onblur field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onfocus(&self) -> Option<Function>

Getter for the onfocus field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onfocus(&self, value: Option<&Function>)

Setter for the onfocus field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onauxclick(&self) -> Option<Function>

Getter for the onauxclick field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onauxclick(&self, value: Option<&Function>)

Setter for the onauxclick field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn oncanplay(&self) -> Option<Function>

Getter for the oncanplay field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_oncanplay(&self, value: Option<&Function>)

Setter for the oncanplay field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn oncanplaythrough(&self) -> Option<Function>

Getter for the oncanplaythrough field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_oncanplaythrough(&self, value: Option<&Function>)

Setter for the oncanplaythrough field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onchange(&self) -> Option<Function>

Getter for the onchange field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onchange(&self, value: Option<&Function>)

Setter for the onchange field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onclick(&self) -> Option<Function>

Getter for the onclick field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onclick(&self, value: Option<&Function>)

Setter for the onclick field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onclose(&self) -> Option<Function>

Getter for the onclose field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onclose(&self, value: Option<&Function>)

Setter for the onclose field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn oncontextmenu(&self) -> Option<Function>

Getter for the oncontextmenu field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_oncontextmenu(&self, value: Option<&Function>)

Setter for the oncontextmenu field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ondblclick(&self) -> Option<Function>

Getter for the ondblclick field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ondblclick(&self, value: Option<&Function>)

Setter for the ondblclick field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ondrag(&self) -> Option<Function>

Getter for the ondrag field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ondrag(&self, value: Option<&Function>)

Setter for the ondrag field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ondragend(&self) -> Option<Function>

Getter for the ondragend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ondragend(&self, value: Option<&Function>)

Setter for the ondragend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ondragenter(&self) -> Option<Function>

Getter for the ondragenter field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ondragenter(&self, value: Option<&Function>)

Setter for the ondragenter field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ondragexit(&self) -> Option<Function>

Getter for the ondragexit field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ondragexit(&self, value: Option<&Function>)

Setter for the ondragexit field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ondragleave(&self) -> Option<Function>

Getter for the ondragleave field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ondragleave(&self, value: Option<&Function>)

Setter for the ondragleave field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ondragover(&self) -> Option<Function>

Getter for the ondragover field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ondragover(&self, value: Option<&Function>)

Setter for the ondragover field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ondragstart(&self) -> Option<Function>

Getter for the ondragstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ondragstart(&self, value: Option<&Function>)

Setter for the ondragstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ondrop(&self) -> Option<Function>

Getter for the ondrop field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ondrop(&self, value: Option<&Function>)

Setter for the ondrop field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ondurationchange(&self) -> Option<Function>

Getter for the ondurationchange field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ondurationchange(&self, value: Option<&Function>)

Setter for the ondurationchange field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onemptied(&self) -> Option<Function>

Getter for the onemptied field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onemptied(&self, value: Option<&Function>)

Setter for the onemptied field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onended(&self) -> Option<Function>

Getter for the onended field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onended(&self, value: Option<&Function>)

Setter for the onended field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn oninput(&self) -> Option<Function>

Getter for the oninput field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_oninput(&self, value: Option<&Function>)

Setter for the oninput field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn oninvalid(&self) -> Option<Function>

Getter for the oninvalid field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_oninvalid(&self, value: Option<&Function>)

Setter for the oninvalid field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onkeydown(&self) -> Option<Function>

Getter for the onkeydown field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onkeydown(&self, value: Option<&Function>)

Setter for the onkeydown field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onkeypress(&self) -> Option<Function>

Getter for the onkeypress field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onkeypress(&self, value: Option<&Function>)

Setter for the onkeypress field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onkeyup(&self) -> Option<Function>

Getter for the onkeyup field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onkeyup(&self, value: Option<&Function>)

Setter for the onkeyup field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onload(&self) -> Option<Function>

Getter for the onload field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onload(&self, value: Option<&Function>)

Setter for the onload field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onloadeddata(&self) -> Option<Function>

Getter for the onloadeddata field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onloadeddata(&self, value: Option<&Function>)

Setter for the onloadeddata field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onloadedmetadata(&self) -> Option<Function>

Getter for the onloadedmetadata field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onloadedmetadata(&self, value: Option<&Function>)

Setter for the onloadedmetadata field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onloadend(&self) -> Option<Function>

Getter for the onloadend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onloadend(&self, value: Option<&Function>)

Setter for the onloadend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onloadstart(&self) -> Option<Function>

Getter for the onloadstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onloadstart(&self, value: Option<&Function>)

Setter for the onloadstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onmousedown(&self) -> Option<Function>

Getter for the onmousedown field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onmousedown(&self, value: Option<&Function>)

Setter for the onmousedown field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onmouseenter(&self) -> Option<Function>

Getter for the onmouseenter field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onmouseenter(&self, value: Option<&Function>)

Setter for the onmouseenter field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onmouseleave(&self) -> Option<Function>

Getter for the onmouseleave field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onmouseleave(&self, value: Option<&Function>)

Setter for the onmouseleave field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onmousemove(&self) -> Option<Function>

Getter for the onmousemove field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onmousemove(&self, value: Option<&Function>)

Setter for the onmousemove field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onmouseout(&self) -> Option<Function>

Getter for the onmouseout field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onmouseout(&self, value: Option<&Function>)

Setter for the onmouseout field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onmouseover(&self) -> Option<Function>

Getter for the onmouseover field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onmouseover(&self, value: Option<&Function>)

Setter for the onmouseover field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onmouseup(&self) -> Option<Function>

Getter for the onmouseup field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onmouseup(&self, value: Option<&Function>)

Setter for the onmouseup field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onwheel(&self) -> Option<Function>

Getter for the onwheel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onwheel(&self, value: Option<&Function>)

Setter for the onwheel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onpause(&self) -> Option<Function>

Getter for the onpause field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onpause(&self, value: Option<&Function>)

Setter for the onpause field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onplay(&self) -> Option<Function>

Getter for the onplay field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onplay(&self, value: Option<&Function>)

Setter for the onplay field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onplaying(&self) -> Option<Function>

Getter for the onplaying field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onplaying(&self, value: Option<&Function>)

Setter for the onplaying field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onprogress(&self) -> Option<Function>

Getter for the onprogress field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onprogress(&self, value: Option<&Function>)

Setter for the onprogress field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onratechange(&self) -> Option<Function>

Getter for the onratechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onratechange(&self, value: Option<&Function>)

Setter for the onratechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onreset(&self) -> Option<Function>

Getter for the onreset field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onreset(&self, value: Option<&Function>)

Setter for the onreset field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onresize(&self) -> Option<Function>

Getter for the onresize field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onresize(&self, value: Option<&Function>)

Setter for the onresize field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onscroll(&self) -> Option<Function>

Getter for the onscroll field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onscroll(&self, value: Option<&Function>)

Setter for the onscroll field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onseeked(&self) -> Option<Function>

Getter for the onseeked field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onseeked(&self, value: Option<&Function>)

Setter for the onseeked field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onseeking(&self) -> Option<Function>

Getter for the onseeking field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onseeking(&self, value: Option<&Function>)

Setter for the onseeking field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onselect(&self) -> Option<Function>

Getter for the onselect field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onselect(&self, value: Option<&Function>)

Setter for the onselect field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onshow(&self) -> Option<Function>

Getter for the onshow field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onshow(&self, value: Option<&Function>)

Setter for the onshow field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onstalled(&self) -> Option<Function>

Getter for the onstalled field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onstalled(&self, value: Option<&Function>)

Setter for the onstalled field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onsubmit(&self) -> Option<Function>

Getter for the onsubmit field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onsubmit(&self, value: Option<&Function>)

Setter for the onsubmit field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onsuspend(&self) -> Option<Function>

Getter for the onsuspend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onsuspend(&self, value: Option<&Function>)

Setter for the onsuspend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ontimeupdate(&self) -> Option<Function>

Getter for the ontimeupdate field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ontimeupdate(&self, value: Option<&Function>)

Setter for the ontimeupdate field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onvolumechange(&self) -> Option<Function>

Getter for the onvolumechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onvolumechange(&self, value: Option<&Function>)

Setter for the onvolumechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onwaiting(&self) -> Option<Function>

Getter for the onwaiting field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onwaiting(&self, value: Option<&Function>)

Setter for the onwaiting field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onselectstart(&self) -> Option<Function>

Getter for the onselectstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onselectstart(&self, value: Option<&Function>)

Setter for the onselectstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ontoggle(&self) -> Option<Function>

Getter for the ontoggle field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ontoggle(&self, value: Option<&Function>)

Setter for the ontoggle field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onpointercancel(&self) -> Option<Function>

Getter for the onpointercancel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onpointercancel(&self, value: Option<&Function>)

Setter for the onpointercancel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onpointerdown(&self) -> Option<Function>

Getter for the onpointerdown field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onpointerdown(&self, value: Option<&Function>)

Setter for the onpointerdown field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onpointerup(&self) -> Option<Function>

Getter for the onpointerup field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onpointerup(&self, value: Option<&Function>)

Setter for the onpointerup field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onpointermove(&self) -> Option<Function>

Getter for the onpointermove field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onpointermove(&self, value: Option<&Function>)

Setter for the onpointermove field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onpointerout(&self) -> Option<Function>

Getter for the onpointerout field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onpointerout(&self, value: Option<&Function>)

Setter for the onpointerout field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onpointerover(&self) -> Option<Function>

Getter for the onpointerover field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onpointerover(&self, value: Option<&Function>)

Setter for the onpointerover field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onpointerenter(&self) -> Option<Function>

Getter for the onpointerenter field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onpointerenter(&self, value: Option<&Function>)

Setter for the onpointerenter field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onpointerleave(&self) -> Option<Function>

Getter for the onpointerleave field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onpointerleave(&self, value: Option<&Function>)

Setter for the onpointerleave field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ongotpointercapture(&self) -> Option<Function>

Getter for the ongotpointercapture field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ongotpointercapture(&self, value: Option<&Function>)

Setter for the ongotpointercapture field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onlostpointercapture(&self) -> Option<Function>

Getter for the onlostpointercapture field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onlostpointercapture(&self, value: Option<&Function>)

Setter for the onlostpointercapture field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onanimationcancel(&self) -> Option<Function>

Getter for the onanimationcancel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onanimationcancel(&self, value: Option<&Function>)

Setter for the onanimationcancel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onanimationend(&self) -> Option<Function>

Getter for the onanimationend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onanimationend(&self, value: Option<&Function>)

Setter for the onanimationend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onanimationiteration(&self) -> Option<Function>

Getter for the onanimationiteration field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onanimationiteration(&self, value: Option<&Function>)

Setter for the onanimationiteration field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onanimationstart(&self) -> Option<Function>

Getter for the onanimationstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onanimationstart(&self, value: Option<&Function>)

Setter for the onanimationstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ontransitioncancel(&self) -> Option<Function>

Getter for the ontransitioncancel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ontransitioncancel(&self, value: Option<&Function>)

Setter for the ontransitioncancel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ontransitionend(&self) -> Option<Function>

Getter for the ontransitionend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ontransitionend(&self, value: Option<&Function>)

Setter for the ontransitionend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ontransitionrun(&self) -> Option<Function>

Getter for the ontransitionrun field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ontransitionrun(&self, value: Option<&Function>)

Setter for the ontransitionrun field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ontransitionstart(&self) -> Option<Function>

Getter for the ontransitionstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ontransitionstart(&self, value: Option<&Function>)

Setter for the ontransitionstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onwebkitanimationend(&self) -> Option<Function>

Getter for the onwebkitanimationend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onwebkitanimationend(&self, value: Option<&Function>)

Setter for the onwebkitanimationend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onwebkitanimationiteration(&self) -> Option<Function>

Getter for the onwebkitanimationiteration field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onwebkitanimationiteration(&self, value: Option<&Function>)

Setter for the onwebkitanimationiteration field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onwebkitanimationstart(&self) -> Option<Function>

Getter for the onwebkitanimationstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onwebkitanimationstart(&self, value: Option<&Function>)

Setter for the onwebkitanimationstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onwebkittransitionend(&self) -> Option<Function>

Getter for the onwebkittransitionend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onwebkittransitionend(&self, value: Option<&Function>)

Setter for the onwebkittransitionend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn onerror(&self) -> Option<Function>

Getter for the onerror field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_onerror(&self, value: Option<&Function>)

Setter for the onerror field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ontouchstart(&self) -> Option<Function>

Getter for the ontouchstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ontouchstart(&self, value: Option<&Function>)

Setter for the ontouchstart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ontouchend(&self) -> Option<Function>

Getter for the ontouchend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ontouchend(&self, value: Option<&Function>)

Setter for the ontouchend field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ontouchmove(&self) -> Option<Function>

Getter for the ontouchmove field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ontouchmove(&self, value: Option<&Function>)

Setter for the ontouchmove field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn ontouchcancel(&self) -> Option<Function>

Getter for the ontouchcancel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn set_ontouchcancel(&self, value: Option<&Function>)

Setter for the ontouchcancel field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn blur(&self) -> Result<(), JsValue>

The blur() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn click(&self)

The click() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement

impl HtmlElement

pub fn focus(&self) -> Result<(), JsValue>

The focus() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlElement