Struct virtual_dom_rs::MouseEvent

#[repr(transparent)]pub struct MouseEvent { /* fields omitted */ }

The MouseEvent class.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

Implementations

impl MouseEvent

pub fn screen_x(&self) -> i32

Getter for the screenX field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn screen_y(&self) -> i32

Getter for the screenY field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn client_x(&self) -> i32

Getter for the clientX field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn client_y(&self) -> i32

Getter for the clientY field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn x(&self) -> i32

Getter for the x field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn y(&self) -> i32

Getter for the y field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn offset_x(&self) -> i32

Getter for the offsetX field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn offset_y(&self) -> i32

Getter for the offsetY field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn ctrl_key(&self) -> bool

Getter for the ctrlKey field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn shift_key(&self) -> bool

Getter for the shiftKey field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn alt_key(&self) -> bool

Getter for the altKey field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn meta_key(&self) -> bool

Getter for the metaKey field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn button(&self) -> i16

Getter for the button field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn buttons(&self) -> u16

Getter for the buttons field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn related_target(&self) -> Option<EventTarget>

Getter for the relatedTarget field of this object.

MDN Documentation

This API requires the following crate features to be activated: EventTarget, MouseEvent

impl MouseEvent

pub fn region(&self) -> Option<String>

Getter for the region field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn movement_x(&self) -> i32

Getter for the movementX field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn movement_y(&self) -> i32

Getter for the movementY field of this object.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn new(type_arg: &str) -> Result<MouseEvent, JsValue>

The new MouseEvent(..) constructor, creating a new instance of MouseEvent.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn get_modifier_state(&self, key_arg: &str) -> bool

The getModifierState() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn init_mouse_event(&self, type_arg: &str)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg(
&self,
type_arg: &str,
can_bubble_arg: bool
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg_and_alt_key_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool,
alt_key_arg: bool
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg_and_alt_key_arg_and_shift_key_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool,
alt_key_arg: bool,
shift_key_arg: bool
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg_and_alt_key_arg_and_shift_key_arg_and_meta_key_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool,
alt_key_arg: bool,
shift_key_arg: bool,
meta_key_arg: bool
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg_and_alt_key_arg_and_shift_key_arg_and_meta_key_arg_and_button_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool,
alt_key_arg: bool,
shift_key_arg: bool,
meta_key_arg: bool,
button_arg: i16
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: MouseEvent, Window

impl MouseEvent

pub fn init_mouse_event_with_can_bubble_arg_and_cancelable_arg_and_view_arg_and_detail_arg_and_screen_x_arg_and_screen_y_arg_and_client_x_arg_and_client_y_arg_and_ctrl_key_arg_and_alt_key_arg_and_shift_key_arg_and_meta_key_arg_and_button_arg_and_related_target_arg(
&self,
type_arg: &str,
can_bubble_arg: bool,
cancelable_arg: bool,
view_arg: Option<&Window>,
detail_arg: i32,
screen_x_arg: i32,
screen_y_arg: i32,
client_x_arg: i32,
client_y_arg: i32,
ctrl_key_arg: bool,
alt_key_arg: bool,
shift_key_arg: bool,
meta_key_arg: bool,
button_arg: i16,
related_target_arg: Option<&EventTarget>
)

The initMouseEvent() method.

MDN Documentation

This API requires the following crate features to be activated: EventTarget, MouseEvent, Window